제발요제발제발료

+++ app/src/main/java/com/example/khyku/User/User.kt
@@ -1,3 +1,3 @@
 package com.example.khyku.User
 
-data class User(val ID:String, var subjects:List<Subject>, var totalTime:Int)
+data class User(val ID:String, var subjects:List<Subject>, var totalTime:Long)
